feat: support decimal scores in grading rules
- Change GradingRule.minScore/maxScore from Int to Float in Prisma schema
- Migration: ALTER COLUMN to DOUBLE PRECISION in PostgreSQL
- API: parseInt → parseFloat in POST/PUT grading-rules handler
- UI: add step="0.1" to all min/max score inputs (edit + add form)

+++ src/app/api/grading-rules/route.ts
@@ -80,8 +80,8 @@ export async function POST(req: NextRequest) {
         const gradingRule = await prisma.gradingRule.create({
             data: {
                 grade,
-                minScore: parseInt(minScore),
-                maxScore: parseInt(maxScore),
+                minScore: parseFloat(minScore),
+                maxScore: parseFloat(maxScore),
                 remark,
                 schoolId,
                 schoolCategory: schoolCategory || null,
@@ -143,8 +143,8 @@ export async function PUT(req: NextRequest) {
             where: { id: existingRule.id },
             data: {
                 grade,
-                minScore: minScore !== undefined ? parseInt(minScore) : undefined,
-                maxScore: maxScore !== undefined ? parseInt(maxScore) : undefined,
+                minScore: minScore !== undefined ? parseFloat(minScore) : undefined,
+                maxScore: maxScore !== undefined ? parseFloat(maxScore) : undefined,
                 remark,
                 // Only update schoolCategory if explicitly provided
                 ...(schoolCategory !== undefined
